// web/lib/types/TenantPolicy.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type TenantPolicy = { deny_latest_tag: boolean, image_allowlist: Array<string>, max_workloads: number | null, max_volumes: number | null, max_volume_mb_total: number | null, allowed_host_ports: Array<number>, require_mfa_admins: boolean, };